[CI] Prevent shellcheck from testing the chef-web-docs submodule
This is code that is periodically updated as part of our documentation
maintenance and publication flow. Since this is not code that is part
of this repository, we shouldn't block our pipeline on any failures
that might be found in it.

Easiest way to do that is to just not check :smile:

Signed-off-by: Christopher Maier <cmaier@chef.io>